+++ codegen.yml
@@ -2,11 +2,22 @@ overwrite: true
 schema: "http://localhost:4000/graphql"
 documents: "graphql/**/*.graphql"
 generates:
-  src/graphql.ts:
+  src/types/cherrytwist-schema.ts:
     plugins:
-      - "typescript"
-      - "typescript-operations"
-      - "typescript-graphql-request"
+      - typescript
+    config:
+      skipTypename: true
+      maybeValue: T | undefined
+  src/graphql.ts:
+    preset: import-types
+    presetConfig:
+      typesPath: ./types/cherrytwist-schema
+      importTypesNamespace: SchemaTypes
+    plugins:
+      - typescript-operations
+      - typescript-graphql-request
     config:
       maybeValue: T | undefined
       rawRequest: true
+      preResolveTypes: true
+      skipTypename: true
