+++ graphql/fragments/challenge-details.graphql
@@ -1,0 +1,13 @@
+fragment ChallengeDetails on Challenge {
+  id
+  name
+  tagset {
+    tags
+    id
+    name
+  }
+  groups {
+    id
+    name
+  }
+}
