+++ graphql/mutations/create/create-challenge.graphql
@@ -1,15 +1,5 @@
 mutation createChallenge($challengeData: ChallengeInput!) {
   createChallenge(challengeData: $challengeData) {
-    name
-    id
-    tagset {
-      tags
-      id
-      name
-    }
-    groups {
-      id
-      name
-    }
+    ...ChallengeDetails
   }
 }
