+++ graphql/queries/challenges.graphql
@@ -1,6 +1,12 @@
-query challenges {
+query challengesBase {
   challenges {
     id
     name
+  }
+}
+
+query challenges {
+  challenges {
+    ...ChallengeDetails
   }
 }
