+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from './CherrytwistClient';
+export * from './types/cherrytwist-schema';
